Criadas novas pastas para entidades e cálculos. Custo total passado agora com também o objeto Petshp.
