+++ TestePratico/Calculos/CalculaCustos.cs
@@ -2,41 +2,21 @@
 using System.Collections.Generic;
 using System.Text;
 
-namespace TestePratico
+namespace TestePratico.Calculos
 {
-    class Petshop
+    class CalculaCustos
     {
-        //Nome do Petshop
-        public string Nome { get; set; }
-        public double Distancia { get; private set; }
-        public double TaxaBanhoDiaDeSemana_P { get; private set; }
-        public double TaxaBanhoDiaDeSemana_G { get; private set; }
-        public double TaxaBanhoFimDeSemana_P { get; private set; }
-        public double TaxaBanhoFimDeSemana_G { get; private set; }
-
-        //Construtor
-        public Petshop(string nome, double distancia,
-            double taxaBanhoDiaDeSemana_P, double taxaBanhoDiaDeSemana_G,
-            double taxaBanhoFimDeSemana_P, double taxaBanhoFimDeSemana_G)
-        {
-            Nome = nome;
-            Distancia = distancia;
-            TaxaBanhoDiaDeSemana_P = taxaBanhoDiaDeSemana_P;
-            TaxaBanhoDiaDeSemana_G = taxaBanhoDiaDeSemana_G;
-            TaxaBanhoFimDeSemana_P = taxaBanhoFimDeSemana_P;
-            TaxaBanhoFimDeSemana_G = taxaBanhoFimDeSemana_G;
-        }
-
         //Calcula custo dos banhos dos cachorros
-        public double CustoTotal(int diaDaSemana, int quantidadeCaesP, int quantidadeCaesG)
+
+        public static double CustoTotal(int diaDaSemana, int quantidadeCaesP, int quantidadeCaesG, Petshop petshop)
         {
             double custoTotal;
 
             //Cálculo feito para pesquisa de segunda a sexta
             //Era um if/else
-            custoTotal = (diaDaSemana > 0 && diaDaSemana < 6)? 
-                (this.TaxaBanhoDiaDeSemana_P * quantidadeCaesP) + (this.TaxaBanhoDiaDeSemana_G * quantidadeCaesG) :
-                (this.TaxaBanhoFimDeSemana_P * quantidadeCaesP) + (this.TaxaBanhoFimDeSemana_G * quantidadeCaesG);
+            custoTotal = (diaDaSemana > 0 && diaDaSemana < 6) ?
+                (petshop.TaxaBanhoDiaDeSemana_P * quantidadeCaesP) + (petshop.TaxaBanhoDiaDeSemana_G * quantidadeCaesG) :
+                (petshop.TaxaBanhoFimDeSemana_P * quantidadeCaesP) + (petshop.TaxaBanhoFimDeSemana_G * quantidadeCaesG);
 
             return custoTotal;
         }
@@ -70,7 +50,7 @@ namespace TestePratico
             else
             {
                 dadosMelhorPreco = Desempate(custos);
-               
+
             }
             return dadosMelhorPreco;
         }
@@ -94,7 +74,5 @@ namespace TestePratico
             return dadosMelhorPreco;
 
         }
-
     }
-    }
-
+}
